requirements.txt
python-telegram-bot>=20.0
pandas>=1.5.0
openpyxl>=3.0.0
xlrd>=2.0.0
xlsxwriter>=3.0.0
